finished working on the errors flagged by the tse
