+++ src/processdata.js
@@ -1,3 +1,4 @@
+
 const processData = (weatherInfo) => {
   const myData = {
     location: weatherInfo.name,
